+++ TestProject/app/src/main/java/net/aquadc/mike/plugin/test/BoxArray.kt
@@ -1,0 +1,7 @@
+package net.aquadc.mike.plugin.test
+
+@Suppress("unused")
+fun boxArrays() {
+    arrayOf(1, 2, 3)
+    Array(5) { it.toFloat() }
+}