+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   <ul>
     <li>Fix interface upcast hints after last fix</li>
     <li>Fix vector path optimizations: don't let numbers stick together when optimizing -0.01 to 0</li>
+    <li>Reporting Kotlin Android Extensions imports as deprecated</li>
   </ul>
 
 ### 0.20