+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     <li>Fix vector path optimizations: don't let numbers stick together when optimizing -0.01 to 0</li>
     <li>Reporting Kotlin Android Extensions imports as deprecated</li>
     <li>Reporting overridden attributes in (layout_margin|padding)(|Left|Top|Right|Bottom|Start|End|Horizontal|Vertical)</li>
+    <li>stop logging error on color state lists when analyzing drawables</li>
   </ul>
 
 ### 0.20